I moved the parse types into their own file.
